Enters the data into the table
Ok , now we can register the information in the table. THe code needs
to be restructured a bit in order to make it cleaner.  Changed the
Display function so that it also display the name of the arrows.

Discovered a bug on runtime when the indicator tokens are instantiated.
Will look this up next