Fixed some indicator issue
When fixing the indicator tokens duplication issue. A naming issue
occured. The new token did not have any name therefore the system could
not find the new position where the tokens need to be flipped. I fixed
that issue by renaming the new token created when we delete the old
token.