Indicators Fix
Tried to fix Yesterda's problem with the indicators not finding the
right positions. Did some progress now it shows more accurately.

Rebulid the code for the indicators from scratch: lookForAvailable It
works better than the previous one. Still need to improve some part as
some locations are not being found.