Making an AI
Started to work on a AI For now it only selects a position randomly.
Changed the Replace Token Script the Start function has been moved to a
Awake fonction. The Ai takes the turn of the  opponent, white. The
issue with the token duplication rose again. Will have to check this
again.