AI finding location
Reminder the Ai is stull random so yeah its not an Ai yet. I fixed the
Indicator selection with  a some delay. It seems it could not find the
token because of the method used to generate the indicators. (It
removes duplication if created, that bugs the unityfunction FIND.
Therefore next time its better to use find with tag)